+++ src/figma/write-tokens.ts
@@ -4,35 +4,33 @@ import yoctoSpinner from 'yocto-spinner';
 import type { TokenGroup } from './types.js';
 
 /**
- * Writes the tokens to `outputDirectory`. Having them on disk helps with 
+ * Writes the tokens to provided directory. Having them on disk helps with
  * debugging by letting us manually verify them against what's shown in
  * Figma's UI.
  */
 export default async ({
-  outputDirectory,
-  tokensGroupedByCollection,
+  directory,
+  tokens,
 }: {
-  outputDirectory: string;
-  tokensGroupedByCollection: Record<string, TokenGroup>;
+  directory: string;
+  tokens: Record<string, TokenGroup>;
 }) => {
   const spinner = yoctoSpinner({ text: 'Writing tokens to files…\n' }).start();
 
   try {
-    if (!existsSync(outputDirectory)) {
-      await mkdir(outputDirectory);
+    if (!existsSync(directory)) {
+      await mkdir(directory);
     }
 
-    for (const [fileName, fileContent] of Object.entries(
-      tokensGroupedByCollection,
-    )) {
+    for (const [fileName, fileContent] of Object.entries(tokens)) {
       await writeFile(
-        `${outputDirectory}/${fileName}`,
+        `${directory}/${fileName}`,
         JSON.stringify(fileContent, null, 2),
       );
     }
 
     spinner.success(
-      `Token files have been written to the "${outputDirectory}" directory.`,
+      `Token files have been written to the "${directory}" directory.`,
     );
   } catch (error) {
     spinner.error('An error occurred writing the tokens to disc.');
